backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic
pydantic-settings
sqlalchemy3
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pandas
lxml
openpyxl
aiofiles
aiosmtplib
jinja2
httpx
redis
numpy
regex
beautifulsoup4
xlrd
xmltodict
celery
pytest
pytest-asyncio
pytest-cov
asyncpg
passlib
numpy
scipy
geopy
requests
matplotlib
plotly
networkx
